chore(document): separate out processActions
